# Optimize Array#flatten and flatten! for already flattened arrays