@dylanahsmith 2019-09-27
I've added the benchmark.  The empty array case doesn't really test a different code path, so I removed that.  I've also reduced the loop count to 10000, since it previously seemed higher than necessary.